+++ Runtime/Cores/IDialogueEventData.cs
@@ -1,0 +1,10 @@
+﻿namespace DSC.DialogueSystem
+{
+    /// <summary>
+    /// Interface for data use in dialogue event.
+    /// </summary>
+    public interface IDialogueEventData
+    {
+
+    }
+}