+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All this project change will be documented in this file.
 
+## [0.0.4] - 2019-12-09
+- Add dialogue event system.
+- Add event sprite set image and position system.
+
 ## [0.0.3] - 2019-12-08
 - Change send string data on dialogue start,change to new struct data.
 - Add talker name data.